+++ meeting_summarizer/__init__.py
@@ -1,5 +1,0 @@
-from meeting_summarizer.my_module import main
-
-__all__ = [
-    'main',
-]