+++ test/unit/test_my_module.py
@@ -1,5 +1,0 @@
-import meeting_summarizer
-
-
-def test_main():
-    meeting_summarizer.main()