+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## Next Release
 
 - Initial release
+- CLI to create a bot and send it to a meeting
+- CLI to summarize a meeting and generate a text file to send to Slack